+++ lib/navigation_wihout_context.dart
@@ -2,4 +2,4 @@ library navigation_wihout_context;
 
 export 'di/register_navigator_get_it_di.dart';
 export 'utils/app_navigator.dart';
-export 'utils/i_app_navigator.dart';
+// export 'utils/i_app_navigator.dart';